WOLO-AI-07- endpoints in one file (app.py) and change translations and suggester to class
WOLO-AI-07- endpoints in one file (app.py) and change translations and suggester to class
