+++ requirements.txt
@@ -3,3 +3,4 @@ transformers~=4.36.2
 torch~=2.1.2
 sentencepiece~=0.1.99
 setfit~=1.0.3
+protobuf
